# BackEnd/docs/swagger.yaml
openapi: 3.0.2
info:
    title: TARgv23ReadingTrackerApi
    version: 1.0.0
    description: API for managing books, users, and ownerships
paths:
    /books:
        get:
            summary: Get all books
            description: Returns a list of all books
            responses:
                '200':
                    description: A list of books
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Book'
    /books/{id}:
        get:
            summary: Get book by ID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            responses:
                '200':
                    description: Details of the book
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Book'
                '404':
                    description: Book not found
        put:
            summary: Update a book
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Book'
            responses:
                '200':
                    description: Updated book details
                '404':
                    description: Book not found
        delete:
            summary: Delete a book by ID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            responses:
                '204':
                    description: Book deleted
                '404':
                    description: Book not found
    /books:
        post:
            summary: Add a new book
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Book'
            responses:
                '201':
                    description: Created book
    /users:
        get:
            summary: Get all users
            responses:
                '200':
                    description: A list of users
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/User'
    /users/{id}:
        get:
            summary: Get user by ID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            responses:
                '200':
                    description: User details
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '404':
                    description: User not found
        put:
            summary: Update a user
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
            responses:
                '200':
                    description: Updated user details
                '404':
                    description: User not found
        delete:
            summary: Delete a user by ID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            responses:
                '204':
                    description: User deleted
                '404':
                    description: User not found
    /users:
        post:
            summary: Add a new user
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
            responses:
                '201':
                    description: Created user
    /ownerships:
        get:
            summary: Get all ownerships
            responses:
                '200':
                    description: A list of ownerships
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Ownership'
    /ownerships/{id}:
        get:
            summary: Get ownership by ID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            responses:
                '200':
                    description: Ownership details
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Ownership'
                '404':
                    description: Ownership not found
        put:
            summary: Update an ownership
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Ownership'
            responses:
                '200':
                    description: Updated ownership details
                '404':
                    description: Ownership not found
        delete:
            summary: Delete an ownership by ID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
            responses:
                '204':
                    description: Ownership deleted
                '404':
                    description: Ownership not found
    /ownerships:
        post:
            summary: Add a new ownership
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Ownership'
            responses:
                '201':
                    description: Created ownership
components:
    schemas:
        Book:
            type: object
            properties:
                BookId:
                    type: integer
                BookName:
                    type: string
                Genre:
                    type: array
                    items:
                        type: string
                ReleaseDate:
                    type: string
                Description:
                    type: string
                ReviewScore:
                    type: integer
                HowManyPages:
                    type: integer
        User:
            type: object
            properties:
                UserId:
                    type: integer
                FirstName:
                    type: string
                LastName:
                    type: string
                UserName:
                    type: string
                Password:
                    type: string
                OwnerShipListID:
                    type: integer
        Ownership:
            type: object
            properties:
                OwnerId:
                    type: integer
                BookId:
                    type: integer
                UserId:
                    type: integer
                Status:
                    type: string